requirements.txt
aiogram>=3.20.0
httpx>=0.28.0
python-dotenv>=1.0.0
Pillow>=10.2.0
